Removing Shoov and adding bash variables
It's unclear to me at this
point whether the shell in which Behat tests run is accessing the
selenium server started in another shell and whether Behat can access
selenium when run with nohup &>/dev/null &.  This remains an area for
further testing.  My preference would be to keep, as much as possible,
commands related to server setup in one step, so that test related steps
only contain meaningful results.  However, if we discover that selenium
does need to be in the same shell as the Behat tests, that can be done.
And if it turns out nohup and sending stdout to /dev/null means Behat
cannot access the selenium server, then we can use GNU parallel.  The
disadvantage with GNU parallel is that we have to guess how long running
the tests might take.